First-pass at the grammar implementation.
This is kind of a huge commit, so checking this in before I finish
writing tests just in case I do something stupid like write a
table-based LL(1) parser manually.

The grammar currently doesn't handle inline tables/arrays, but that's
just a matter of adding more states. The implementation maintains a
stack so all the bits that parsing a CFL requires are there.