Fix issues with array-of-tables within tables.
 * Correctly append them when the path contains an array.
 * Emit an error when a declared array interferes with an inline one.

This now almost passes all of the burntsushi tests -- the only errors
now are due to output inconsistencies (toml-test wants inline array
emit in a different format -- libtoml2 makes no distinction between an
inline array containing objects and an array of tables).

(or, it didn't previously, but now that's technically tracked. So I
guess it's totally fixable!)