Fix a mis-interpretation of mine about array-of-tables.
When encountering a subfield in an array-of-tables declaration, e.g.

    [[album.songs]]

When resolving the subfield a new entry should not be appended to the
albums array if there are entries -- it should only append to an empty
array. If there's an element in the "album" array it should select the
last one. This allows declarations like

    [[albums]]
    [[album.songs]]
    [[album.songs]]
    [[albums]]
    [[album.songs]]

To correctly create the top-level array with 2 elements rather than 5
(since the [[album.songs]] should modify the most recent [[album]]
entry rather than creating a new one).

With this, all of toml-test is satistfied. Huzzah.